feat: bundle runtime code
