+++ packages/rspack/src/builtin-plugin/css-extract/index.ts
@@ -11,7 +11,7 @@ import { PLUGIN_NAME } from "./utils";
 export * from "./loader";
 
 const DEFAULT_FILENAME = "[name].css";
-const LOADER_PATH = join(__dirname, "css-extract-loader.js");
+const LOADER_PATH = join(__dirname, "cssExtractLoader.js");
 
 export type { CssExtractRspackLoaderOptions } from "./loader";
